requirements.txt
slackclient>=2.0.0
certifi
beautifulsoup4
requests